change type field rate from double to BigDecimal

+++ src/main/java/com/aleks/currency_exchange/view/ExchangeRateView.java
@@ -2,18 +2,19 @@ package com.aleks.currency_exchange.view;
 
 import com.aleks.currency_exchange.model.Currency;
 
+import java.math.BigDecimal;
+
 public class ExchangeRateView {
 
     private int id;
     private Currency baseCurrency;
     private Currency targetCurrency;
-    private double rate;
+    private BigDecimal rate;
 
-    public ExchangeRateView(int id, Currency baseCurrency, Currency targetCurrency, double rate) {
+    public ExchangeRateView(int id, Currency baseCurrency, Currency targetCurrency, BigDecimal rate) {
         this.id = id;
         this.baseCurrency = baseCurrency;
         this.targetCurrency = targetCurrency;
         this.rate = rate;
     }
-
 }
